refactor: 커뮤니티 상세페이지, 후기 상세페이지를 구분하기 위하여 파일 구조 리팩토링
