+++ src/components/Community/WriteDetail/DetailContainer/DetailContainer.jsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import DetailHeader from '../DetailHeader/DetailHeader';
-import Comments from '../../Comments/Comments';
+import Comments from '../../../Comments/Comments';
 import detailContainerStyle from './DetailContainer.style';
 
 export default function DetailContainer() {
