+++ src/components/Community/WriteDetail/DetailContainer/DetailContainer.style.jsx
