+++ src/components/Community/WriteDetail/DetailHeader/DetailHeader.jsx
@@ -1,5 +1,5 @@
 /** @jsxImportSource @emotion/react */
-import MockupData from '../../Community/CommunityPopularPosts/MockupData';
+import MockupData from '../../CommunityPopularPosts/MockupData';
 import {
   headerContainer,
   mainInfoStyle,
@@ -8,8 +8,8 @@ import {
   reactionContainerStyle,
   hrStyle,
 } from './DetailHeader.style';
-import Hashtags from '../../Hashtag/Hashtag';
-import Career from '../../Career/Career';
+import Hashtags from '../../../Hashtag/Hashtag';
+import Career from '../../../Career/Career';
 
 export default function DetailHeader() {
   const [{ img, title, username, career, date, body, hashtags, likes, comments, views }] = MockupData.slice(0, 1);
