+++ src/components/Community/WriteDetail/DetailHeader/DetailHeader.style.jsx
