+++ src/pages/Community/CommunityDetail.jsx
@@ -1,5 +1,5 @@
 import BackgroundRectangle from '../../components/BackgroundRectangle/BackgroundRectangle';
-import DetailContainer from '../../components/Detail/DetailContainer/DetailContainer';
+import DetailContainer from '../../components/Community/WriteDetail/DetailContainer/DetailContainer';
 import Nav from '../../components/Nav';
 
 export default function Community() {
